backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-google-vertexai==2.0.1
langchain-core==0.3.15
google-cloud-aiplatform==1.72.0
googlemaps==4.10.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
motor==3.6.0
pymongo==4.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
Jinja2==3.1.4
aiohttp==3.10.10
cachetools==5.5.0
